Split block entity into 2
Split the block entity obstacle into upper and lower blocks.
This allows for a gap in between. The vertical length of the
blocks go outside the window height to account for changes in the
block gap's location. The plan is for the block gap to be moved
during the game by changing the y transform of both the upper
and lower blocks.